Fix compatibility bugs with Numeric and without any numpy (r3b demo).
git-svn-id: https://svn.code.sf.net/p/auto-07p/code/trunk/auto/07p@1877 ee9e63dd-6847-0410-88cd-b7174c777ff7

+++ python/Points.py
@@ -388,7 +388,7 @@ class Point(object):
             try:
                 ct = kw['coordtype']
             except KeyError:
-                self.coordtype = float
+                self.coordtype = float64
             else:
                 try:
                     self.coordtype = _num_equivtype[ct]
